Move the repository class
