--- lib/docker/repository.rb
@@ -1,24 +1,0 @@
-require 'deploy/repository'
-require 'docker/release/tag_dockerrun'
-
-module Docker
-  class Repository < ::Deploy::Repository
-    private
-
-    def tag_dockerrun!
-      Release::TagDockerrun.new(@tag).call
-    end
-
-    def version!
-      super
-      tag_dockerrun!
-    end
-
-    def commit!
-      puts 'Committing version.txt and Dockerrun.aws.json...'
-      unless system('git add public/version.txt Dockerrun.aws.json') && system("git commit -m \"#{commit_message}\" ")
-        fail 'Failed to commit.'
-      end
-    end
-  end
-end
